+++ SimTracker/SiStripDigitizer/plugins/SiHitDigitizer.h
@@ -51,7 +51,7 @@ class SiHitDigitizer {
   }
 
   void processHit(const PSimHit*, const StripGeomDetUnit&, GlobalVector,float,
-		  std::vector<double>&, size_t&, size_t&,
+		  std::vector<float>&, size_t&, size_t&,
 		  const TrackerTopology *tTopo);
   
  private:
